feat: Add Version type
Add an `Epoch` type to track the optional epoch of a package version.
Add a `Pkgver` type to track the pkgver of a package version.
Add a `Pkgrel` type to track the optional pkgrel of a package version.
Add a `Version` type to track epoch, pkgrel and pkgrel of a
package version (relying on `Epoch`, `Pkgver` and `Pkgrel`).
Add `Version::vercmp()` as an implementation of libalpm/pacman's
`vercmp`.