chore(license): Relicense project as Apache-2.0 OR MIT
Rely on new REUSE.toml and relicense the project under the terms of the
Apache-2.0 and MIT (so it may be used as "Apache-2.0 OR MIT").
Relicense any GFDL-1.3-or-later as CC-BY-SA-4.0 to standardize on a
single documentation license.

Remove all attribution, as it is now tracked centrally in REUSE.toml.

Fixes https://gitlab.archlinux.org/archlinux/alpm/alpm-types/-/issues/39

Signed-off-by: David Runge <dvzrv@archlinux.org>