feat: Add justfile and integration for checks and tests
Add justfile to provide all check and test targets.
Add an environment file to track all necessary Arch Linux packages for
this project.
Add a lychee configuration to ignore links that either point to
imaginary files or to documentation targets that do not yet exist.
Extend the reuse configuration to set licensing information for the
new files.

Signed-off-by: David Runge <dvzrv@archlinux.org>