test(buildinfo): test with snapshots
Utilize `insta` crate to do snapshot testing for the alpm-buildinfo
implementation. This helps with visualizing and debugging the test output.

The snapshots are saved under the `test/snapshots` directory with their
respective test name. When a new test is added or the snapshot has been
changed, it should be investigated why that happened or the snapshots
should be _accepted_ with using the `cargo insta` command line tool.

Simply run `cargo insta review` to see the differences between the last
snapshot and press `a` or `r` to approve/reject the changes
respectively.

Signed-off-by: Orhun Parmaksız <orhunparmaksiz@gmail.com>