refactor(buildinfo): create a helper function for guessing Schema
This commit adds a new `from_contents` function to the `Schema` for
deserializing the contents of a BUILDINFO file and guessing the schema
version from the "format" field.

This allows the `--schema` option in the `validate` function to be
optional as well, as we can guess the version from the file itself.

Signed-off-by: Orhun Parmaksız <orhunparmaksiz@gmail.com>
Signed-off-by: Arne Beer <contact@arne.beer>