test(buildinfo): Add integration tests for buildinfo v2
Extend the existing tests to cover both the buildinfo V1 and V2
formats by calling the `validate` subcommand.

The file is moved to `integration.rs` while at it, since it no longer
only covers the v1 buildinfo and it was deemed convenient enough to have
both formats in the same file.

Signed-off-by: Arne Beer <contact@arne.beer>