feat(alpm-parser): Use winnow for custom-ini parsing
Replace the manual parsing logic that splits keys and values with a
proper winnow parsing logic.

This also fixes an existing bug in the old manual parsing logic that
allowed a delimiter to not be surrounded by spaces.

This is however required by the spec to differentiate between spaces
that belong to the delimiter and spaces that belong to the string.

Signed-off-by: Arne Beer <contact@arne.beer>