feat(mdbook): Add just recipes for an mdbook integration
Add `build-book` recipe to build the mdbook setup and assemble
documentation in a dedicated output directory.

Add `serve-book` recipe to serve the book with miniserve (as we are
staging the rust docs alongside we can't make use of the mdbook builtin
as it would delete the rust docs).

Add `watch-book` recipe to watch for changed files using watchexec and
rebuild using mdbook (as we are staging the rust docs alongside we can't
make use of the mdbook builtin as it would delete the rust docs).

Fixes: https://gitlab.archlinux.org/archlinux/alpm/alpm/-/issues/76
Signed-off-by: David Runge <dvzrv@archlinux.org>