feat(srcinfo): Add srcinfo parser
Add the first parser pass of a SRCINFO file.

The goal of this is to get a line-based representation of a SRCINFO file
that already ensures several things:

- pkgname and pkgbase sections may only contain keywords that're valid
  in the scope of their section.
- The values for each keyword have the proper structure/format and will
  be converted into their correct Rust typed equivalent.

The resulting output is rather raw and not suitable for user consumption
yet. However, the line based it allows us to do proper linting in the
next step!

Signed-off-by: Arne Beer <contact@arne.beer>