feat(srcinfo): SourceInfo struct representation
Add the actual struct representation of the raw SRCINFO file.
This format is still fairly raw as one needs to do a lot of value
merging to get to the actual merged result for a single package for a
single architecture.

During this step, both errors and linter warnings may be returned, which
need to be handled by the user.
The idea is to return all errors in a single go to make it easier for
users to fix everything at once instead of having to run the command
multiple times.

Signed-off-by: Arne Beer <contact@arne.beer>