refactor(srcinfo): Move command line functions away from lib.rs
First up: Please forgive me that I didn't incorporate these changes into
the MR history. The MR is so big at this point in time that this
would've easily taken me another hour to do so and I would've messed up
somewhere. Anyhow, this MR is a perfectly fine encapsultated change
in its own and kind of deserves its own message, as it derivates from
the API design in our other parser.

While designing the API surface, I noticed that it really doesn't make a
lot of sense to expose the functions used for the commandline as
top-level functions in the library. They're just not that well-suited
for library users.

They should use the helper functions on the SourceInfo struct instead.

Signed-off-by: Arne Beer <contact@arne.beer>