refactor(pkginfo): Clean up API surface
Before this commit, the functions called by the binary were at the top
level of the library.

Those functions however are highly related to the cli module and
executable logic. They should be hidden as well as possible for the
library consumer, as this isn't how they'll interact with this.

Signed-off-by: Arne Beer <contact@arne.beer>