fix(tests): Replace testdir with tempfile
testdir has some issues where tempdirs might be reused in the scope of a
single testing thread.
Due to this behavior, we encountered flaky tests in both the pkginfo and
buildinfo crate.

Replace testdir with tempfile, which is already in our dependency tree
due to the dev-scripts anyway. That library is more popular,
well-tested and doesn't run into these issues as it generates a truly
random directory/file in the user's tmpfs, which is automatically
cleaned up as soon as the variable goes out of scope.

Signed-off-by: Arne Beer <contact@arne.beer>