refactor(types): Properly type PackageRelease version data
Before this commit, we parsed the PackageRelease, but kept the original
string around for some reason.

This has been fixed and PackageRelease now just wraps a major version
with an optional minor version without.

Additionally the previous PackageRelease comparison was a simple string
comparison, which was obviously wrong.
The new code uses uses the typed PackageRelease version information to
perform proper version comparison.

Resolves #153

Signed-off-by: Arne Beer <contact@arne.beer>