feat(cargo): Use a custom target runner for x86_64-unknown-linux-gnu
Use a custom target runner script that enables distinguishing between
running tests on the host and tests in a containerized environment.
Distinguishes between containerized and non-containerized tests by
relying on the test module name, which must contain `containerized` to
be considered a test to be containerized.
Add podman to the list of development dependencies.

Signed-off-by: David Runge <dvzrv@archlinux.org>