refactor: Change architectures to Vec
Previously, architectures have been a Vec on both Package and
PackageBase. The reason for this was to enforce the invariant
of no duplicate architectures on a type-level.

The downside of this was that the original order in which those
architectures were read was lost.

As both `makepkg` and `devtools` use that order to determine the order
in which properties are written to a `.SRCINFO` file, I deemed it
necessary to change this back to a Vec so that we're able to write
idempotent SRCINFO files.

The main purpose of being able to being idempotent is that this makes
testing and ensuring backwards compatibility **much** easier.

Signed-off-by: Arne Beer <contact@arne.beer>