feat!: Use `FullVersion`, not `Version` in `InstalledPackage`
The installed package is always represented by an alpm-package-version
of the form _full_ or _full with epoch_.
To make the type more robust, rely on the more strict `FullVersion`
instead of `Version`.

Change `InstalledPackage::new` to be infallible, as it not longer needs
to return an `Error` (and also never did in fact...).

Signed-off-by: David Runge <dvzrv@archlinux.org>