feat: Add alpm-soname auto-deps
This feature reads the input directory of a built package and generates
provide and depend entries by reading ELF files in the package and
extracting their sonames, and the sonames they depend on.

This functions the same as pacman`s autodeps feature introduced in
v6.1.0. Though arch does not use autodeps and have patched it out of
their pacman build.

A simple command line tool is included to help play with the feature.
It prints provides and depends as pkginfo entries. Though this is just
for display purposes and is not meant to actually be used to create
.PKGINFO files.

To fully implement the autodeps feature, depend entries must only be
added if the depend is currently satisfied. As we currently have no
tooling for installing packages, there is an escape hatch via the
`--test-satisfied-command`. This flag calls a command to check if
a depend is currently satisfied. This can be used to call pacman like so

```
alpm-soname auto-deps --test-satisfied-command 'pacman -T --' \
    tests/test_files/pacman
```

This is a clunky hack but is currently useful for playing around with.

This isn't an issue on the library side as the check will be done on
a higher level by the caller. So we don't need to worry about it there.

The interface for the command is quite different to the existing
`get-provisions` and `get-dependencies` commands as it is one command
that handles both provides and depends. I think the existing commands
should merge to be more similar to auto-deps.

I'm also not sure if the name autodeps should be used as it's the name
of the pacman feature. It seems everything in this project is trying to
be a bit generic with naming, instead of just adopting the pacman
terminoligy.

The crate `lddtree` is used to resolve ELF dependencies. It seems to
work perfectly but it`s a small project. Maybe this should be
implemented by us in the future if we need to?

Testing is implementing by reusing the simple meson project used to test
`find_provision`s and `find_dependencies`. There is also a test against
a cut down pacman package containing just the pacman binary and
libalpm.so.

This feature differs somewhat with the existing terminoligy in
alpm-soname. `depends` and `provides` is used over `dependencies` and
`provisions`. I think this terminoligy is better as the shorer width
means the cursor drifts to the right less. These are also the terms
used in pacman, `provider` or `provide` are tossed around way more often
and feel less clunky than `provision`. I would prefer it if we changed
the terminoligy elsewhere in alpm-soname.

Signed-off-by: morganamilo <morganamilo@archlinux.org>