feat: Introduce alpm-lint framework
Introduction of the alpm-lint linting framework.

This commit adds too much to be explained in this commit message, but
two documents, ARCHITECTURE.md and CONTRIBUTING.md have been created
that give an architectural overview and an explanation on how to add new
lints respectively.

I think it would be a good idea if you read through these documents
before exploring the crate. If there's anything unclear, the documents
then apparently need some adjustment.

Signed-off-by: Arne Beer <contact@arne.beer>