feat: Add alpm-lint documentation website
This commit adds a static site to serve documentation about all
available alpm-lint rules.

The project uses `zola` in combination with the `linkita` theme.

The website mostly consists of the index page, where one can view,
search for and filter all available rules.

The page is built by building the `alpm-lint` binary and extracting all
metadata in the form of JSON. That metadata is then pulled by zola and
used to template the index page.

The stack used for this is:

- Zola with Tera templating
- Tailwind for styling (which ships with linkita)
- Javascript for search and other interactive behavior.
- SCSS for a few classes that're required for convenient javascript
  interaction (bundled via zola)

Signed-off-by: Arne Beer <contact@arne.beer>