fix: Duplicate makepkg option
Removes a duplicate 'debug' value from the PackageOption enum.

While at it, use `strum::VariantNames` to implement VARIANTS in a less
error-prone way.

In theory, we could also use strum's Display impl to handle the
serialization of variants, but that sadly clashes with the existing
pattern of using `Display` to convert the types into their
PKGBUILD/SRCINFO "!flag" representation.

Signed-off-by: Arne Beer <contact@arne.beer>