fix: Trailing whitespace handling
This fix changes the way whitespace handling is done in the SourceInfo
parser.

Before this commit, leading/trailing whitespaces were handled **inside**
the Package/PackageProperty parsers, which lead to issues in the case
`eof` was hit.

If an `eof` after whitespaces was encountered, the surrounding `repeat`
loop should terminate via backtrack. But due to the backtrack, the
cursor was reset to **before** the whitespaces.

This is a typical case of parsers trying to handle content that doesn't
belong to them.

To work around this, whitespace handling has moved to their caller
functions. Both Package and PackageProperty now expect that there're no
leading whitespaces!

This is ensured by enforcing the pattern of:
- Trim leading whitespaces before the first loop.
- Trim trailing whitespaces **after** each successful loop via.
    `terminated(RawPackage::parser, multispace0)`

Signed-off-by: Arne Beer <contact@arne.beer>