feat: Add trailing separator to dirs collected by `relative_files`
By default, a `PathBuf` does not add a trailing separator (e.g. "/" or
"\") to directory paths.
However, this is widely used in the ALPM context to help distinguish
directories from regular files.

Adding a trailing separator character to directory paths is
unproblematic, as e.g. joining of `PathBuf` still works as before.

Signed-off-by: David Runge <dvzrv@archlinux.org>