refactor: Cleanup soname modules, dependencies and feature flags
The `commands` module is not supposed to be usable outside of the
binary. Hence, it was made a private module for the binary only.

The motivation for this is that library users shouldn't pull
libraries that they don't need. Specifically, the `clap` and
`serde_json` dependency.
Until now users had to explicitly disable default features for them
to not be included. With this change, this is now done by default
and the `cli` feature flag needs to be manually enabled.

The `cli` module has been kept part of the library for now as this
is needed for us to generate the man pages.

Also, to further keep CLI logic out of the library, a new error enum
has been introduced to the `commands` module, which wraps the crate's
`Error` type and contains error variants that only appear during CLI
handling.

Additionally, while working on this I noticed that the logging is
exposed all over the place and the library doesn't even build with
default features disabled. So I made the `log` crate required for
now.

If we want to make logging a separate feature flag, this needs to be
done separately.

Signed-off-by: Arne Beer <contact@arne.beer>