feat(lint): Add `UnknownArchitecture` lint
Ensures that all alpm-architecture names specified in the
SRCINFO are known. Any string passing the alpm-architecture
requirements is a valid architecture. However, in most cases,
only a limited set of architecture names is used in practice.
This lint rule checks whether any uncommon architecture names
are used in the SRCINFO which often indicates a typo.

Warnings emitted by this lint rule can be safely ignored if
the specified architecture name is correct, but simply uncommon.

Signed-off-by: Jagoda Ślązak <jslazak@jslazak.com>