fix: Enable creating completions and manpages for `alpm-db-desc`
The alpm-db crate is the first crate in the workspace that is setup to
have multiple executable.
This comes at the disadvantage of our rust-script and justfile based
generation of manpages and shell completions to no longer work.

As a stop-gap solution, we allow setting a specific module for a
CLI, until we have decided how we want to integrate manpage and shell
completion integration in the future.

Signed-off-by: David Runge <dvzrv@archlinux.org>