fix: Ensure that sections in `alpm-db-desc` adhere to the specification
This fixes a few inconsistencies in the initial implementation of the
`alpm-db-desc` parser and writer and adds further tests to ensure
correctness going forward.

The initial implementation assumed, that the `%URL%` and `%DESC%`
sections could be unset in input and would also omit the respective
sections in output, if no value is provided for them.
However, this does not reflect the specification and current behavior of
existing tooling, as the section headers are always written (even if the
values are the empty string) and should be expected on input.

Further, the `%SIZE%` section could not be omitted in input.
In case the `%SIZE%` section does not exist in input, it should lead to
a value of `0` for that section though.

The `%REASON%` section, if omitted in input, would lead to a
value of `None` but should instead result in the assignment of
`PackageInstallReason::Explicit`.
If the section holds a value of `PackageInstallReason::Explicit` (i.e.
0), it should be omitted from output, but previously wasn't.

The `%VALIDATION%` section was implemented as a list value, but instead
should be a single value.

Fixes: https://gitlab.archlinux.org/archlinux/alpm/alpm/-/issues/239
Signed-off-by: David Runge <dvzrv@archlinux.org>