refactor: Split up the `alpm_types::relation` module further
To make the addition of new functionality easier and have fewer very
long files, split up the `alpm_types::relation` module further.

The lower-level relation types `Group`, `OptionalDependency` and
`PackageRelation` are grouped in a `base` module and the ELF/soname
specific types `SharedLibraryPrefix`, `Soname`, `SonameV1`, `SonameV2`,
`VersionOrSoname` are moved to a `soname` module.

The `composite` module holds the higher-level `RelationOrSoname` type
which is a composition of types from the above modules.

Signed-off-by: David Runge <dvzrv@archlinux.org>